backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic==2.5.3
sqlalchemy==2.0.25
requests==2.31.0
firebase-admin==6.3.0 